+++ src/Alert/Alert.js
@@ -17,10 +17,12 @@ export class Alert extends Component {
 
   render() {
       const {
+          buttonProps,
           type,
           link,
+          linkProps,
           linkText,
-          dismissable,
+          dismissible,
           children,
           className,
           ...props
@@ -29,13 +31,14 @@ export class Alert extends Component {
           <div>
               {this.state.isActive && (
                   <div
-                      className={`fd-alert${dismissable ? ' fd-alert--dismissible' : ''}${
+                      className={`fd-alert${dismissible ? ' fd-alert--dismissible' : ''}${
                           type ? ' fd-alert--' + type : ''
                       }${className ? ' ' + className : ''}`}
                       role='alert'
                       {...props}>
-                      {dismissable && (
+                      {dismissible && (
                           <button
+                              {...buttonProps}
                               aria-controls='j2ALl423'
                               aria-label='Close'
                               className='fd-alert__close'
@@ -43,7 +46,10 @@ export class Alert extends Component {
                       )}
                       {children}
                       {link && (
-                          <a className='fd-link' href={link}>
+                          <a
+                              {...linkProps}
+                              className='fd-link'
+                              href={link}>
                               {linkText}{' '}
                               <span className='sap-icon--arrow-right sap-icon--s' />
                           </a>
@@ -56,9 +62,11 @@ export class Alert extends Component {
 }
 
 Alert.propTypes = {
+    buttonProps: PropTypes.object,
     className: PropTypes.string,
-    dismissable: PropTypes.bool,
+    dismissible: PropTypes.bool,
     link: PropTypes.string,
+    linkProps: PropTypes.object,
     linkText: PropTypes.string,
     type: PropTypes.oneOf(['', 'warning', 'error', 'success', 'information'])
 };